+++ praxis-core/src/Language/Praxis/PRA/Lemmas.hs
@@ -1,0 +1,6 @@
+{-# LANGUAGE QuasiQuotes #-}
+
+module Language.Praxis.PRA.Lemmas () where
+
+import Language.Praxis.PRA.PrimitiveRecursion
+import Language.Praxis.PRA.Tactic.Quote